Open quick pallet for folders that starting on the same characters
If you would have 3 folders like
 - foobar
 - folder
 - folder_with_longer_name

When you type `fo` and tab it will open quick pallet as before.
Previously IOpener would append a / to the end but it's not unique. In
bash's tab completion it does not add a slash and I don't think we
should either.

Now it will open a quick panel with folder and folder_with_longer_name
and you can choose.